requirements.txt
python-dotenv==1.0.0
PyAudio==0.2.13
pydub==0.25.1
numpy==1.24.3
gTTS==2.3.1
langdetect==1.0.9
opencv-python==4.6.0.66
scipy==1.9.3
mediapipe==0.10.1
joblib==1.2.0
openai==1.13.3
openai-whisper==20231117
pvporcupine==2.1.4
pvrecorder==1.1.1
telepota==1.0
spotdl==4.2.0
psutil==5.9.4
pycaw==20230407
pyserial==3.5